[Agent] Fixed test to pass on windows (#16922)
[Agent] Fixed test to pass on windows (#16922)

(cherry picked from commit cfcae9da83f8e2817683c0f18fd45a7c5c0e9c70)
