+++ x-pack/agent/pkg/agent/operation/monitoring_test.go
@@ -2,8 +2,6 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-// +build linux darwin
-
 package operation
 
 import (
